+++ src/a2x_pack_sound.p.h
@@ -31,7 +31,7 @@ extern void a_music_stop(void);
 
 extern ASound* a_sfx_fromFile(const char* Path);
 #define a_sfx_fromData(Data) a_sfx__fromData(Data, sizeof(Data))
-extern ASound* a_sfx__fromData(const uint16_t* Data, int Size);
+extern ASound* a_sfx__fromData(const uint8_t* Data, int Size);
 
 extern void a_sfx_play(ASound* Sfx);
